Making the output of a Model custom flexible
It would be nice that the models had a special method, that if defined
was called for the string representation.

Usually __unicode__ is defined in a verbose way for debugging purposes.
Other times you may point to a model that has many fields represented in
the __unicode__ method, but for the user you just want to show one.